Refactor: PR 리뷰 반영 : 오타 수정

+++ src/main/java/org/programmers/staybb/domain/room/Option.java
@@ -1,10 +1,9 @@
 package org.programmers.staybb.domain.room;
 
 import com.sun.istack.NotNull;
-import lombok.Getter;
-
-import javax.persistence.Embeddable;
 import java.util.Objects;
+import javax.persistence.Embeddable;
+import lombok.Getter;
 
 @Embeddable
 @Getter
@@ -15,12 +14,12 @@ public class Option {
     @NotNull
     private int bedroomNum;
     @NotNull
-    private float bathroomNum;
+    private double bathroomNum;
 
     protected Option() {
     }
 
-    public Option(int bedNum, int bedroomNum, int bathroomNum) {
+    public Option(int bedNum, int bedroomNum, double bathroomNum) {
         this.bedNum = bedNum;
         this.bedroomNum = bedroomNum;
         this.bathroomNum = bathroomNum;
